refactor: divided main.go into stummer.go and utils.go
